this is a temporary fix to update the shpc parser.
The gh:// parser treats a recipe extension (e.g., salad) as a tag, and then the
release from the GitHub version as the digest. This sort of assumes that
every release has the same tags, or the tags that are desired are represented
in the repository. Afaik Binoc/cuppa does not do any custom parsing of the artifacts
to discover tags. So this PR fixes a bug that the release versions are being used as
tags, and the name of the repository as the digest.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>